+++ IsabellaCateringWebApp/models/Models/tblPaymentsModel.cs
@@ -9,9 +9,9 @@ namespace IsabellaCateringWebApp.Models.Models
     {
         public int paymentID { get; set; }
         public int bookingID { get; set; }
-        public float amountDue { get; set; }
-        public float amount { get; set; }
-        public float remainingBalance {  get; set; }
+        public double amountDue { get; set; }
+        public double amount { get; set; }
+        public double remainingBalance {  get; set; }
         public int transactionNum { get; set; }
         public string paymentType { get; set; }
         public string paymentStatus { get; set; }
